mass icon/name changes
Added the ability to choose your own icon for 'check' properties,
as well as changing the name of the property. This is then updated
accross the board to all other days. Still needs a lot of refactoring.
Only works with one property/icon at a time for now, will work on
multiple changes at once in the future